% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_angle.R
\name{get_best_angle}
\alias{get_best_angle}
\title{Get best angle for plotting the tree}
\usage{
get_best_angle(las)
}
\arguments{
\item{las}{An object of class LAS [lidR::readLAS()].}
}
\value{
Returns a list containing the model object with the required parameters and model_type used.
}
\description{
Calculates the minimum oriented bounding box using the
rotating calipers algorithm and extracts the angle
}
\examples{
lasfile <- system.file("extdata", "tree_c2.laz", package = "rTLsDeep")
las <- lidR::readLAS(lasfile)

(get_best_angle(las))

}
