% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_angle.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Rotating calipers algorithm}
\usage{
getMinBBox(xy)
}
\arguments{
\item{xy}{A matrix of xy values from which to calculate the minimum oriented
bounding box.}
}
\description{
Calculates the minimum oriented bounding box using the
rotating calipers algorithm.
Credits go to Daniel Wollschlaeger <https://github.com/ramnathv>
}
