% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTLsDeep-package.R
\docType{package}
\name{rTLsDeep}
\alias{rTLsDeep-package}
\alias{rTLsDeep}
\title{rTLsDeep: Tools for post-hurricane individual tree-level damage classification}
\description{
The rTLsDeep package provides tools for:
\itemize{
  \item rotating and deriving 2D images from TLS 3D point clouds
  \item calibrating and validating convolutional neural network (CNN) architectures
  \item predicting post-hurricane damage severity at the individual tree level
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carlos-alberto-silva/rTLsDeep}
  \item Report bugs at \url{https://github.com/carlos-alberto-silva/rTLsDeep/issues}
}

}
\author{
\strong{Maintainer}: Caio Hamamura \email{caiohamamura@gmail.com} [contributor]

Authors:
\itemize{
  \item Carine Klauberg \email{carine_klauberg@hotmail.com}
  \item Ricardo Dalagnol \email{r.valbuena@bangor.ac.uk} [copyright holder]
  \item Matheus Ferreira [contributor]
  \item Jason Vogel [contributor]
  \item Carlos Alberto Silva \email{carlos_engflorestal@outlook.com} [copyright holder]
}

}
\keyword{internal}
