% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_angle.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Rotating calipers algorithm}
\usage{
getMinBBox(hull)
}
\arguments{
\item{hull}{A matrix of xy values from a convex hull from which
will calculate the minimum oriented bounding box.}
}
\description{
Calculates the minimum oriented bounding box using the
rotating calipers algorithm.
}
