% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTLsDeep-package.R
\docType{package}
\name{rTLsDeep-package}
\alias{rTLsDeep}
\alias{rTLsDeep-package}
\title{rTLsDeep: Post-Hurricane Damage Severity Classification from TLS and AI}
\description{
Terrestrial laser scanning (TLS) data processing and post-hurricane damage severity classification at the individual tree level using deep Learning. Further details were published in Klauberg et al. (2023) \doi{10.3390/rs15041165}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carlos-alberto-silva/rTLsDeep}
  \item Report bugs at \url{https://github.com/carlos-alberto-silva/rTLsDeep/issues}
}

}
\author{
\strong{Maintainer}: Caio Hamamura \email{caiohamamura@gmail.com} [contributor]

Authors:
\itemize{
  \item Carine Klauberg \email{carine_klauberg@hotmail.com}
  \item Ricardo Dalagnol \email{r.valbuena@bangor.ac.uk} [copyright holder]
  \item Matheus Ferreira [contributor]
  \item Jason Vogel [contributor]
  \item Carlos Alberto Silva \email{carlos_engflorestal@outlook.com} [copyright holder]
}

}
\keyword{internal}
