% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTLsDeep.R
\docType{package}
\name{rTLsDeep}
\alias{rTLsDeep}
\title{rTLsDeep: Set of tools for post-hurricane individual tree-level damage classification from terrestrial laser scanning and deep learning.}
\description{
The rTLSDeep package provides options for:
i) rotating and deriving 2D images from TLS 3D point clouds
ii) calibrating and validating convolutional neural network (CNN) architectures and 
iii) predicting post-hurricane damage severity at the individual tree level
}
