% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlsdeep_setup.R
\name{rtlsdeep_setup}
\alias{rtlsdeep_setup}
\title{Setup python and tensorflow environment}
\usage{
rtlsdeep_setup(python_version = "3.12")
}
\arguments{
\item{python_version}{Character string with Python version to install if not found. Default: "3.12".}
}
\value{
Nothing
}
\description{
Setup python and tensorflow environment
}
