% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlsdeep_setup.R
\name{rtlsdeep_setup}
\alias{rtlsdeep_setup}
\title{Setup Python and TensorFlow environment}
\usage{
rtlsdeep_setup()
}
\value{
Invisible TRUE on success.
}
\description{
Installs Python and TensorFlow if they are not already available.
This function is intended for interactive use only.
}
