% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlsdeep_setup.R
\name{rtlsdeep_setup}
\alias{rtlsdeep_setup}
\title{Setup Python and TensorFlow environment}
\usage{
rtlsdeep_setup()
}
\value{
Invisible TRUE on success.
}
\description{
Installs Python and TensorFlow if they are not already available.
This function is intended for interactive use only.
}
\examples{
\dontrun{
tensorflow_dir = NA
model_type = "simple"
train_image_files_path = system.file('extdata', 'train', package='rTLsDeep')
test_image_files_path = system.file('extdata', 'validation', package='rTLsDeep')
img_width <- 256
img_height <- 256
class_list_train = unique(list.files(train_image_files_path))
class_list_test = unique(list.files(test_image_files_path))
lr_rate = 0.0001
target_size <- c(img_width, img_height)
channels = 4

rtlsdeep_setup()
}
}
